Persitance is working
:-D

+++ src/slimWallet/slimWallet/Model/CardModel.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.ObjectModel;
+﻿using System;
+using System.Collections.ObjectModel;
 using System.IO;
 using System.Threading.Tasks;
 using GalaSoft.MvvmLight;
@@ -43,8 +44,11 @@ namespace slimWallet.Model
         {
             _database = new Database();
             _fileRepository = new FileRepository();
-            List = new ObservableCollection<Card>(_database.GetItemsAsync().Result);
         }
+
+        public async Task Init() {
+            if (List == null) List = new ObservableCollection<Card>(await _database.GetItemsAsync());
+        } 
 
         public async Task SaveAsync(Card card)
         {
@@ -62,17 +66,19 @@ namespace slimWallet.Model
             await _fileRepository.DeleteAsync(card.BackImage);
         }
 
+        public async Task<Stream> ReadAsync(string fileName) => await _fileRepository.ReadAsync(fileName);
+
         public async Task SaveFileAsync(Card selected, Stream stream, bool front)
         {
             var path = await _fileRepository.SaveAsync(stream);
             stream.Dispose();
 
             if (front) {
-                await _fileRepository.DeleteAsync(selected.FrontImage);
+                if(selected.FrontImage != null) await _fileRepository.DeleteAsync(selected.FrontImage);
                 selected.FrontImage = path;
             }
             else {
-                await _fileRepository.DeleteAsync(selected.FrontImage);
+                if (selected.BackImage != null) await _fileRepository.DeleteAsync(selected.FrontImage);
                 selected.BackImage = path;
             }
         }
